header section style added in easy-bank.css file.
This css file having navigation bar and related elements. It sets up color variables for various shades of gray, blue, green, and white. The CSS ensures a clean, responsive design: the navigation bar sticks to the top of the page as users scroll, and it includes a toggle button for small screens. The button transforms into a cross when clicked, revealing a menu with links to different sections of the website. There's also a stylish gradient button for inviting users. The code prioritizes accessibility and user experience, making it easy for visitors to navigate through the site.