Main class added in easy-bank.html file.
This code represents the main content section of a webpage. Inside a 'main' element, there is a 'section' with a class of home and an id of home. This section likely serves as the homepage banner for Easybank website. It contains an image displayed within a box, showcasing Easybank's services. Below the image, there is a heading with the text 'Next generation digital banking', indicating the main message of the page. A paragraph below the heading provides additional details about Easybank's services, describing it as a comprehensive platform for various financial activities. Finally, there is a button labeled 'Request Invite', which users can click to take action, presumably to sign up or learn more about Easybank's services.