Main class  style added in easy-bank.css file.
In this CSS code, specific styles for the main content of a webpage are defined. The 'main' element's overflow is set to hidden, ensuring that any content overflowing its container is hidden from view. The '.home' section has a light gray background and extra padding at the bottom. Inside this section, there is an 'home-img-box' containing an image styled to cover its container and align at the bottom. The 'home-title' class sets the font size, color, and alignment for a large heading in the center of the section, while the 'home-text' class adjusts the font size, color, and line spacing for a paragraph of text beneath the heading. These styles collectively create a visually appealing and well-structured homepage layout for a website.