Service class added in easy-bank.html file.
In the ,a section called 'service' is created, likely representing a segment of a website that highlights the benefits of Easybank. Inside this section, there is a heading with the text 'Why choose Easybank?' and a paragraph explaining how Easybank utilizes Open Banking to transform users' bank accounts into comprehensive financial hubs, offering enhanced control over their finances. Below the text, there are four service cards, each containing an icon representing a feature (such as online banking, simple budgeting, fast onboarding, and Open API), a title describing the feature, and a description explaining the benefits. These elements collectively present Easybank's key services and advantages to potential users in a clear and straightforward manner.