artical class style and media query added in easy-bank.css file.
In this code styles specify the appearance of different sections on the webpage. The 'blog' section has a light gray background and padding for spacing. Each article card within the blog section is styled with a white background, rounded corners, and a hover effect on the article titles.  different screen sizes to create a responsive and visually appealing website layout. For small screens (less than 768 pixels wide), adjustments are made to the font sizes, padding, and layout of the home, service, and blog sections to ensure readability and proper spacing. On medium-sized screens (768 to 1023 pixels wide), the font sizes are increased, and the layout of the service and blog sections is optimized with two columns. For larger screens (1024 to 1199 pixels wide), further adjustments are made to the header, navigation bar, home section, service section, and blog section to enhance the overall design. On extra-large screens (1200 pixels and wider), additional padding is added around the content to improve spacing and aesthetics. These responsive design techniques ensure a seamless user experience across various devices and screen sizes.