Footer class added in easy-bank.html file.
The footer section of the website is defined. The footer contains a logo, social media links (Facebook, YouTube, Twitter, Pinterest, Instagram), navigation links for 'About Us', 'Contact', and 'article', as well as links for 'Careers', 'Support', and 'Privacy Policy'. There is also a 'Request Invite' button and a copyright notice indicating that the website is owned by Easybank. Additionally, there's an attribution line indicating the coder's name (Dinesh A). The code also includes Ionicons library for icon rendering. These elements together create the footer section of the website.